webapi/webapi-contactsmanager-w3c-tests/contactsmanager/Contact_attributes.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, Kang <kangx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>ContactsManager Test: Contact_attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/2012/sysapps/contacts-manager-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>
var contact;

setup(function () {
  var data = new Date();
  var url = new ContactField({
    types: ['work'],
    preferred: true,
    value: 'http://www.intel.com/'
  });
  var emailAddr = new ContactField({
    types: ['work'],
    preferred: true,
    value: 'tester@sample.com'
  });
  var address = new ContactAddress({
    types: ['work'],
    preferred: true,
    countryName: 'china'
  });
  var mobilePhone = new ContactTelField({
    types: ['home'],
    preferred: true,
    value: '+34698765432'
  });
  var instantAddr = new ContactField({
    types: ['home'],
    preferred: true,
    value: 'msn:tester@sample.com'
  });
  var contactName = new ContactName({
    givenNames: ['John'],
    familyNames: ['Doe']
  });
  contact = new Contact({
    name: contactName,
    phoneNumbers: [mobilePhone],
    urls: [url],
    emails: [emailAddr],
    photos: ["test"],
    categories: ['swimmer,biker'],
    addresses: [address],
    organizations: ["Bubba Gump Shrimp Co."],
    jobTitles: ["Shrimp Man"],
    birthday: data,
    notes: ["I am proficient in Tiger-Crane Style"],
    impp: [instantAddr],
    anniversary: data,
    gender: 'male'
  });
});

test(function () {
  assert_true("name" in contact, "The Contact.name exists");
  assert_equals(typeof contact.name, "object", "The type of Contact.name");
}, "Check if the attribute Contact.name exists and typeof object");

test(function () {
  assert_true("emails" in contact, "The Contact.emails exists");
  assert_equals(typeof contact.emails, "object", "The type of Contact.emails");
}, "Check if the attribute Contact.emails exists and typeof object");

test(function () {
  assert_true("photos" in contact, "The Contact.photos exists");
  assert_equals(typeof contact.photos, "object", "The type of Contact.photos");
}, "Check if the attribute Contact.photos exists and typeof object");

test(function () {
  assert_true("urls" in contact, "The Contact.urls exists");
  assert_equals(typeof contact.urls, "object", "The type of Contact.urls");
}, "Check if the attribute Contact.urls exists and typeof object");

test(function () {
  assert_true("categories" in contact, "The Contact.categories exists");
  assert_equals(typeof contact.categories, "object", "The type of Contact.categories");
}, "Check if the attribute Contact.categories exists and typeof object");

test(function () {
  assert_true("addresses" in contact, "The Contact.addresses exists");
  assert_equals(typeof contact.addresses, "object", "The type of Contact.addresses");
}, "Check if the attribute Contact.addresses exists and typeof object");

test(function () {
  assert_true("phoneNumbers" in contact, "The Contact.phoneNumbers exists");
  assert_equals(typeof contact.phoneNumbers, "object", "The type of Contact.phoneNumbers");
}, "Check if the attribute Contact.phoneNumbers exists and typeof object");

test(function () {
  assert_true("organizations" in contact, "The Contact.organizations exists");
  assert_equals(typeof contact.organizations, "object", "The type of Contact.organizations");
}, "Check if the attribute Contact.organizations exists and typeof object");

test(function () {
  assert_true("jobTitles" in contact, "The Contact.jobTitles exists");
  assert_equals(typeof contact.jobTitles, "object", "The type of Contact.jobTitles");
}, "Check if the attribute Contact.jobTitles exists and typeof object");

test(function () {
  assert_true("birthday" in contact, "The Contact.birthday exists");
  assert_equals(typeof contact.birthday, "object", "The type of Contact.birthday");
}, "Check if the attribute Contact.birthday exists and typeof object");

test(function () {
  assert_true("notes" in contact, "The Contact.notes exists");
  assert_equals(typeof contact.notes, "object", "The type of Contact.notes");
}, "Check if the attribute Contact.notes exists and typeof object");

test(function () {
  assert_true("impp" in contact, "The Contact.impp exists");
  assert_equals(typeof contact.impp, "object", "The type of Contact.impp");
}, "Check if the attribute Contact.impp exists and typeof object");

test(function () {
  assert_true("anniversary" in contact, "The Contact.anniversary exists");
  assert_equals(typeof contact.anniversary, "object", "The type of Contact.anniversary");
}, "Check if the attribute Contact.anniversary exists and typeof object");

test(function () {
  assert_true("gender" in contact, "The Contact.gender exists");
  assert_equals(typeof contact.gender, "string", "The type of Contact.gender");
}, "Check if the attribute Contact.gender exists and typeof string");
</script>
